+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Add icon
+
+### Changed
+
+-   support more functions and classes in `predefinition.py`
+
 ## [0.0.3] - 2025-06-25
 
 ### Added